+++ src/compile.task.ts
@@ -1,12 +1,12 @@
 import {Task} from './task/task.js';
 import {compileSourceDirectory} from './compile/compileSourceDirectory.js';
-import {loadBuildConfigs} from './build/buildConfig.js';
+import {loadGroConfig} from './config/config.js';
 
 export const task: Task = {
 	description: 'compiles all files to the build directory',
 	run: async ({log}) => {
 		// TODO how to do this?
-		const dev = process.env.NODE_ENV === 'development';
-		await compileSourceDirectory(await loadBuildConfigs(), dev, log);
+		const dev = process.env.NODE_ENV !== 'production';
+		await compileSourceDirectory(await loadGroConfig(), dev, log);
 	},
 };