+++ src/dev.task.ts
@@ -2,21 +2,21 @@ import {Task} from './task/task.js';
 import {Filer} from './build/Filer.js';
 import {createDevServer} from './devServer/devServer.js';
 import {createDefaultCompiler} from './compile/defaultCompiler.js';
-import {paths, toBuildOutDir} from './paths.js';
-import {loadBuildConfigs, loadPrimaryBuildConfig} from './build/buildConfig.js';
+import {paths, toBuildOutPath} from './paths.js';
 import {loadTsconfig, toEcmaScriptTarget} from './compile/tsHelpers.js';
+import {loadGroConfig} from './config/config.js';
 
 export const task: Task = {
 	description: 'start development server',
 	run: async ({log}): Promise<void> => {
-		const buildConfigs = await loadBuildConfigs();
-		const primaryBuildConfig = await loadPrimaryBuildConfig();
-		const buildConfigToServe =
-			primaryBuildConfig.platform === 'browser'
-				? primaryBuildConfig
-				: buildConfigs.find((c) => c.platform === 'browser') || primaryBuildConfig;
-		const buildOutDirToServe = toBuildOutDir(true, buildConfigToServe.name, '');
+		const config = await loadGroConfig();
+		// TODO should this be `findServedBuildConfig`? or should this be a property on the config itself?
+		// maybe that gets added by a normalization step?
+		const buildConfigToServe = config.primaryBrowserBuildConfig ?? config.primaryNodeBuildConfig;
+		const buildOutDirToServe = toBuildOutPath(true, buildConfigToServe.name, '');
 
+		// TODO probably replace these with the Gro config values
+		// maybe the default config reads these from tsconfig?
 		const tsconfig = loadTsconfig(log);
 		const target = toEcmaScriptTarget(tsconfig.compilerOptions?.target);
 		const sourceMap = tsconfig.compilerOptions?.sourceMap ?? true;
@@ -25,7 +25,7 @@ export const task: Task = {
 			compiler: createDefaultCompiler(),
 			compiledDirs: [paths.source],
 			servedDirs: [buildOutDirToServe],
-			buildConfigs,
+			buildConfigs: config.builds,
 			target,
 			sourceMap,
 		});