+++ src/dist.task.ts
@@ -1,10 +1,10 @@
 import {copy} from './fs/nodeFs.js';
 import {Task} from './task/task.js';
-import {paths, toBuildOutDir} from './paths.js';
+import {paths, toBuildOutPath} from './paths.js';
 import {isTestBuildFile, isTestBuildArtifact} from './oki/testModule.js';
 import {printPath} from './utils/print.js';
 import {cleanDist} from './project/clean.js';
-import {loadBuildConfigs} from './build/buildConfig.js';
+import {loadGroConfig} from './config/config.js';
 
 export const isDistFile = (path: string): boolean =>
 	!isTestBuildFile(path) && !isTestBuildArtifact(path);
@@ -15,15 +15,13 @@ export const task: Task = {
 		await cleanDist(log);
 
 		// This reads the `dist` flag on the build configs to help construct the final dist directory.
-		// See the docs at `./project/buildConfig.md`.
-		const dev = process.env.NODE_ENV === 'development';
-		const buildConfigs = await loadBuildConfigs();
-		const buildConfigsForDist = buildConfigs.some((c) => c.dist)
-			? buildConfigs.filter((c) => c.dist)
-			: buildConfigs;
+		// See the docs at `./docs/config.md`.
+		const dev = process.env.NODE_ENV !== 'production';
+		const config = await loadGroConfig();
+		const buildConfigsForDist = config.builds.filter((b) => b.dist);
 		await Promise.all(
 			buildConfigsForDist.map((buildConfig) => {
-				const buildOutDir = toBuildOutDir(dev, buildConfig.name);
+				const buildOutDir = toBuildOutPath(dev, buildConfig.name);
 				const distOutDir =
 					buildConfigsForDist.length === 1 ? paths.dist : `${paths.dist}${buildConfig.name}`;
 				log.info(`copying ${printPath(buildOutDir)} to ${printPath(distOutDir)}`);